+++ src/components/ImportLoadStep.vue
@@ -59,13 +59,13 @@
       :items="formatItems"
       v-model="form.format"
     />
-    <div class="prose prose-sm">
+    <div class="prose-sm prose">
       <p v-if="isMilx">
         Basic support for importing MilX layers from
         <a href="https://www.map.army/">map.army</a>
       </p>
     </div>
-    <p v-if="isMilx" class="prose prose-sm">
+    <p v-if="isMilx" class="prose-sm prose">
       Please note that the import functionality is experimental.
     </p>