+++ src/components/ImportMilxStep.vue
@@ -1,29 +1,33 @@
 <template>
-  <form @submit.prevent="onLoad" class="mt-4 space-y-6">
-    <div class="prose prose-sm">
-      <p v-if="isMilx">
-        Basic support for importing MilX layers from
-        <a href="https://www.map.army/">map.army</a>
-      </p>
-    </div>
-    <section>
-      <div v-for="layer in data" :key="layer.name">
-        <p>{{ layer.name }}</p>
-        <ul class="ml-4 space-y-1.5 font-mono">
-          <li v-for="f in layer.features" class="flex items-center">
-            <MilSymbol :sidc="f.properties.sidc" :size="20" /><span class="ml-2">{{
-              f.properties.name || f.properties.sidc
-            }}</span>
-          </li>
-        </ul>
+  <div class="">
+    <form @submit.prevent="onLoad" class="mt-4 flex max-h-[80vh] flex-col">
+      <div class="flex-auto overflow-auto">
+        <div class="prose-sm prose">
+          <p v-if="isMilx">
+            Basic support for importing MilX layers from
+            <a href="https://www.map.army/">map.army</a>
+          </p>
+        </div>
+        <section class="p-1.5">
+          <!--          <p class="text-base font-medium">Options</p>-->
+          <!--          <InputCheckbox label="Include units" />-->
+          <SimpleSelect
+            label="Parent unit"
+            :items="rootUnitItems"
+            v-model="parentUnitId"
+          />
+        </section>
+        <section class="mt-4">
+          <ImportMilxStepTable :layers="data" v-model:selected="selectedUnits" />
+        </section>
       </div>
-    </section>
 
-    <footer class="flex items-center justify-end space-x-2 pt-4">
-      <BaseButton type="submit" primary small disabled>Import</BaseButton>
-      <BaseButton small @click="emit('cancel')">Cancel</BaseButton>
-    </footer>
-  </form>
+      <footer class="flex flex-shrink-0 items-center justify-end space-x-2 pt-4">
+        <BaseButton type="submit" primary small>Import</BaseButton>
+        <BaseButton small @click="emit('cancel')">Cancel</BaseButton>
+      </footer>
+    </form>
+  </div>
 </template>
 
 <script setup lang="ts">
@@ -34,7 +38,14 @@ import type { ImportFormat, ImportSettings } from "@/types/convert";
 import { useNotifications } from "@/composables/notifications";
 import { useImportStore } from "@/stores/importExportStore";
 import { MilxImportedLayer, useScenarioImport } from "@/composables/scenarioImport";
-import MilSymbol from "@/components/MilSymbol.vue";
+
+import { injectStrict, nanoid } from "@/utils";
+import { activeScenarioKey } from "@/components/injects";
+import { NUnit } from "@/types/internalModels";
+import ImportMilxStepTable from "@/components/ImportMilxStepTable.vue";
+import { Point } from "geojson";
+import SimpleSelect from "@/components/SimpleSelect.vue";
+import { SelectItem } from "@/components/types";
 
 interface Props {
   data: MilxImportedLayer[];
@@ -42,9 +53,9 @@ interface Props {
 
 const props = defineProps<Props>();
 const emit = defineEmits(["cancel", "loaded"]);
-
+const { unitActions, store: scnStore } = injectStrict(activeScenarioKey);
 const store = useImportStore();
-
+const { state } = scnStore;
 interface Form extends ImportSettings {
   format: ImportFormat;
 }
@@ -60,10 +71,41 @@ const form = ref<Form>({
 
 const { focusId } = useFocusOnMount(undefined, 150);
 const { send } = useNotifications();
+const selectedUnits = ref<(string | number)[]>([]);
 
 const isMilx = computed(() => form.value.format === "milx");
 const { importMilxString } = useScenarioImport();
+
+const sides = computed(() => {
+  return state.sides.map((id) => state.sideMap[id]);
+});
+
+const rootUnitItems = computed((): SelectItem[] => {
+  return Object.values(state.sideGroupMap)
+    .map((value) => value.subUnits)
+    .flat()
+    .map((e) => state.unitMap[e])
+    .map((u) => ({ label: u.name, value: u.id }));
+});
+
+const parentUnitId = ref(rootUnitItems.value[0].value as string);
+
 async function onLoad(e: Event) {
-  const { format } = form.value;
+  const features = props.data.map((l) => l.features).flat();
+
+  const units: NUnit[] = features
+    .filter((e) => selectedUnits.value.includes(e.id!))
+    .map((f) => {
+      return {
+        id: nanoid(),
+        name: f.properties.name || "",
+        sidc: f.properties.sidc,
+        subUnits: [],
+        _pid: "",
+        location: (f.geometry as Point).coordinates,
+      };
+    });
+  units.forEach((unit) => unitActions.addUnit(unit, parentUnitId.value));
+  emit("loaded");
 }
 </script>