+++ src/modules/storymode/StoryModeContent.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
     <p class="fixed bottom-2 right-2 z-40 border bg-white p-2">{{ sIndex }}</p>
-    <div class="prose prose-sm p-4" v-html="renderedContent"></div>
+    <div class="prose-sm prose p-4" v-html="renderedContent"></div>
   </div>
 </template>