+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Basic MilX (.milxlyz) import from map.army
+
+October 2022
+
 - Add letter based SIDC to number SIDC converter to symbol picker
 - Grid cell copy/paste
 - Grid edit mode keyboard navigation